Only handle GA authentication on the defined redirect URI
This plugin was trying to perform oAuth authentication with Google
anytime it received a URL containing the query string '?code='. In
addition to breaking any other oauth integrations on a site, this
seems like just a bad idea in general.

This commit checks to make sure that the oauth handshake only happens on
the page defined as the ga_redirect_uri in the plugin settings -
ideally, this would be an admin page to prevent abuse.